+++ phpBB2/admin/page_header_admin.php
@@ -64,9 +64,8 @@ $template->set_filenames(array(
 	'header' => 'admin/page_header.tpl')
 );
 
-// Format Timezone. We are unable to use array_pop here, because of PHP3 compatibility
-$l_timezone = explode('.', $board_config['board_timezone']);
-$l_timezone = (count($l_timezone) > 1 && $l_timezone[count($l_timezone)-1] != 0) ? $lang[sprintf('%.1f', $board_config['board_timezone'])] : $lang[number_format($board_config['board_timezone'])];
+// Format timezone, including custom legacy offsets not present in a language pack.
+$l_timezone = phpbb_timezone_label($board_config['board_timezone']);
 
 //
 // The following assigns all _common_ variables that may be used at any point