+++ phpBB2/includes/page_header.php
@@ -566,9 +566,8 @@ foreach ($nav_links as $nav_item => $nav_array)
 	}
 }
 
-// Format Timezone. We are unable to use array_pop here, because of PHP3 compatibility
-$l_timezone = explode('.', $board_config['board_timezone']);
-$l_timezone = (count($l_timezone) > 1 && $l_timezone[count($l_timezone)-1] != 0) ? $lang[sprintf('%.1f', $board_config['board_timezone'])] : $lang[number_format($board_config['board_timezone'])];
+// Format timezone, including custom legacy offsets not present in a language pack.
+$l_timezone = phpbb_timezone_label($board_config['board_timezone']);
 
 /* CrackerTracker IP Range Scanner */
 $marknow = (isset($HTTP_GET_VARS['marknow']) && is_scalar($HTTP_GET_VARS['marknow'])) ? (string) $HTTP_GET_VARS['marknow'] : '';